@@ -1,4 +1,4 @@
-import 'package:auto_route/auto_route.dart';
+/*import 'package:auto_route/auto_route.dart';
 import 'package:estudo_app/app/app_controller.dart';
 import 'package:estudo_app/config/router/base_router.gr.dart';
 import 'package:estudo_app/src/ui/utils/constants/app_colors.dart';
@@ -13,6 +13,10 @@ class AppWidget extends StatefulWidget {
 }
 
 class _AppWidgetState extends BaseState<AppWidget, AppController> {
+  GlobalKey navigatorKey = GlobalKey<NavigatorState>();
+  GlobalKey key = LabeledGlobalKey<State<StatefulWidget>>('key');
+  // GlobalKey key = GlobalKey();
+
   @override
   void initState() {
     super.initState();
@@ -20,6 +24,37 @@ class _AppWidgetState extends BaseState<AppWidget, AppController> {
 
   @override
   Widget build(BuildContext context) {
+    return OKToast(
+      child: MaterialApp(
+        key: key,
+        builder: ExtendedNavigator.builder(
+          navigatorKey: key,
+          router: BaseRouter(),
+          name: NavigatorHelper.baseRouterName,
+          builder: (context, extendedNav) {
+            final data = MediaQuery.of(context);
+            return MediaQuery(
+              data: data.copyWith(
+                textScaleFactor: 1,
+              ),
+              child: Theme(
+                data: ThemeData(
+                  primaryColor: AppColors.primaryColor,
+                ),
+                child: extendedNav,
+              ),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}*/
+import 'package:flutter/material.dart';
+
+class AppWidget extends StatelessWidget {
+  // final _baseRouter = BaseRouter();
+  /*Widget build(BuildContext context) {
     return OKToast(
       child: MaterialApp(
         builder: ExtendedNavigator.builder(
@@ -41,6 +76,18 @@ class _AppWidgetState extends BaseState<AppWidget, AppController> {
           },
         ),
       ),
+    );
+  }*/
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Home'),
+        ),
+      ),
+      debugShowCheckedModeBanner: false,
     );
   }
 }
