@@ -1,11 +1,10 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:injectable/injectable.dart';
 
 @injectable
 class RetryHelper {
   Future<void> testRetry({
-    @required int timeToRetry,
-    @required int maximumAttempts,
+    int timeToRetry,
+    int maximumAttempts,
   }) async {
     var currentAttempts = 0;
     await Future.doWhile(() async {
